requirements.txt
fp-NGFW-SMC-python==0.7.0b20
PyYAML==5.3.1
requests~=2.23.0